fix 699
